requirements.txt
facenet-pytorch>=2.6.0,<3.0
scikit-learn>=1.4,<1.6
pandas>=2.2,<3
matplotlib>=3.8,<3.10
joblib>=1.3,<2
Flask>=3,<4
python-dotenv>=1.0,<2
gunicorn>=21,<22
